fix: ボタンの状態管理
