+++ Assets/Scripts/StudyWatch/InitStory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6de68f795788bd346a4d503ed5792198
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
