+++ Assets/Scripts/StudyWatch/ListButton.cs
@@ -2,10 +2,19 @@ using UnityEngine;
 
 public class ListButton : MonoBehaviour
 {
+    [SerializeField] GameObject _closeButton;
+    [SerializeField] GameObject _starButton;
 
-    private void Start()
+    public void SetCloseButton()
     {
+        _closeButton.gameObject.SetActive(true);
+        _starButton.gameObject.SetActive(false);
     }
 
+    public void SetStarButton()
+    {
+        _closeButton.gameObject.SetActive(false);
+        _starButton.gameObject.SetActive(true);
+    }
 
 }
